Catch errors on calibration file generation and continue
This is mainly related to when a calibration interval is not well defined
in the input files. I think it is better to continue with the rest
of the input files rather than crash here but maybe we need
to handle it better. Maybe we should yield a Result object?